+++ src/main/kotlin/com/autonomousapps/tasks/FindDeclaredProcsTask.kt
@@ -125,8 +125,8 @@ abstract class FindDeclaredProcsTask : DefaultTask() {
       if (procs != null) artifact to procs else null
     }.flatMap { (artifact, procs) ->
       procs.mapNotNull { procName ->
-        inMemoryCache.procs[procName] ?: procFor(artifact, procName, classLoader!!).also { proc ->
-          proc?.let { inMemoryCache.procs.putIfAbsent(procName, it) }
+        inMemoryCache.proc(procName) ?: procFor(artifact, procName, classLoader!!).also { proc ->
+          proc?.let { inMemoryCache.procs(procName, it) }
         }
       }
     }