+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Dependency Analysis Plugin Changelog
 # Version 0.60.0 (unreleased)
 * [Fixed] Does not union global and local rules as expected.
 ([Issue 273](https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/issues/273))
+* [Fixed] In-memory cache now has an eviction policy.
+([Issue 274](https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/issues/274))
 
 # Version 0.59.0
 * [New] Kotlin stdlib now a dependency bundle by default, improving ergonomics for Kotlin users.